add ConfiguredMethodRewriteRule
